examples/mem0-performance-comparison/pyproject.toml
[project]
name = "mem0-benchmark"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "mem0ai>=1.0.1",
    "fastembed",
]
